+++ docs/README.md
@@ -3,7 +3,7 @@
 - -[x] 컴퓨터측 난수 생성 기능
 - -[x] 게임 시작을 안내하는 기능
 - -[x] 숫자를 입력받는 기능
-- -[ ] 숫자야구 게임룰에 맞게 난수와 숫자 비교 기능
+- -[x] 숫자야구 게임룰에 맞게 난수와 숫자 비교 기능
 - -[ ] 결과 출력 기능
 - -[ ] 재시작 기능
 - -[ ] 종료 기능
